kbuild: fix DTC-native include path for dts/ directory
For the pre-processor to include sources below dts/ directory,
-I$(srctree)/dts/src is specified as an include path.  On the other
hand, for the DTC-native, -i $(srctree)/dts/src/$(SRCARCH) is
currently specified.

Theoretically, the pre-processor style (#include) and DTC-native
style (/include/) should work with the same include paths.

Currently, there are no users for the latter, so we can fix it safely.

Signed-off-by: Masahiro Yamada <yamada.masahiro@socionext.com>